fix: rename Chip to BaseChip and IconChip to Chip
